backend/pyproject.toml
[project]
name = "agent-backend"
version = "0.1.0"
description = "Agent — AI Harness Backend (FastAPI + LangGraph)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
    "pyjwt>=2.9.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends._legacy:_Backend"
